fix: 그룹 채팅방 기본 이미지 생성 로직 버그 수정
- ID 3번 이상의 그룹 채팅방에서 500 에러 발생하는 문제 해결
- 기본 이미지 인덱스 계산 공식을 (id-1)%count+1에서 id%count+1로 수정

+++ src/main/java/com/sonsminpark/auratalkback/domain/chat/service/ChatRoomImageServiceImpl.java
@@ -37,7 +37,7 @@ public class ChatRoomImageServiceImpl implements ChatRoomImageService {
         log.debug("채팅방 기본 이미지 생성 - 채팅방 ID: {}", chatRoomId);
 
         try {
-            int index = Math.toIntExact(chatRoomId % DEFAULT_GROUP_IMAGE_COUNT) + 1;
+            int index = (int) (chatRoomId % DEFAULT_GROUP_IMAGE_COUNT) + 1;
 
             // 범위 체크
             if (index < 1 || index > DEFAULT_GROUP_IMAGE_COUNT) {
